Shrink Twitter sidebar, restricting font size
I would've used JavaScript to dynamically calculate the height, but
there's no simple way to get the width of an element on insertion. The
alternative (which is what is in use) is to fix the width of Twitter and
calculate the ratio manually.

Once an aspect ratio property/unit is finalized, I can replace the
complex calculation with something along the lines of `calc(9ar / 16)`,
and replace the `22rem` with `minmax(auto, 22rem)`, which would let the
Twitter component shrink as appropriate.