Prompt user for thread ID if one is not present in URL
I've also updated the dependencies, allowing my to add cssnano in the
same plugin stack for PostCSS (there was a bug that prevented this
previously). As a result of these updates, I had to slightly modify the
YouTube component as something (incorrectly) strips away an empty string
check.

To avoid a circular dependency (for rollup), I've extracted the element
definitions to a separate file.

Because loading data (necessary to check if the thread ID is valid) is
inherently asynchronous, I've had to modify a large portion of the
initialization code to make is async. Hopefully this refactor makes it
easier to read, though!