Remove eslint & remnants, lint gulpfile, inline CSS in JS
CSS in JS will reduce the number of server requests, which should
improve speed slightly. That's on top of the improved compression of a
larger file.

Current loading requires only a few requests:
- index.html
- interface.bundle.js
- modules.bundle.js (includes all components' CSS)

The service worker can also be loaded optionally, and a request to the
API is necessary to load in all data.