Switch to table layout for Events section
This allows for dynamically sized columns. It's relatively difficult to
get a fixed table header in place _along with_ the section header, so at
least for now the table header will scroll with the text.

Once subgrid is released, this component can be rewritten, and should be
significantly cleaner and easier to follow.