Conversion to custom syntax
We now use babel-plugin-custom-element for all components. which
provides for a clean, easy to understand syntax and dynamic updates.

I've also introduced Rollup into the mix, which combines the JavaScript
and PostCSS into a single file (for what's needed initially). If the
user is a host, Rollup does its own code splitting.

Some PostCSS plugins have been removed, as they either weren't used or
may be confusing to other contributors. All that's left is simple
iteration, inlined SVGs, and stage 2+ options from cssdb (along with a
couple stage 1 options).

This commit does not fully complete the transition to the new syntax
from Lit Element & Lit HTML, as I still need to handle the addition of
another layer of Shadow DOM in the Sections, as well as the dnyamic
reordering of Sections and Events (which will now be handled by a small
helper function, rather than a full vDOM).

The other major change is switching from HTML to Pug, which is (to me)
easier to read quickly and find what you want. It also allows for some
JavaScript iterpolation, so I can set the thtme color, title, and
favicon sizes in one place at the top of the file, generating the
appropriate elements at build time.